Cross-Level Puzzles and More Sound
Since that bug guy is not there to give the player the fake items, as of right now, I just give the items to the player when the player presses E. So, the method is set up, it will just have to be called once that guy is integrated into the game. Then each suitor house can be used once for trading. By the way, I used my 2 playground scenes for all of my integrating and testing.

Also, I added sound effects for the movable block, pressure plates, the interact button, and the camera shift.